Remove administrator check
We want to be able to run the agent on boxes where we
do not have sudo access. We cannot use a deb or rpm on
these servers and need a self-contained logdna agent
install.

The build.sh script can be used to package up node and the
agent bits together in a single archive that can be copied
onto the server that it needs to run on.